Issue #11732: Refactor Frequency into support-base component
